+++ CHANGELOG.md
@@ -31,6 +31,7 @@
     - Changed Metric and Report API
     - Added Text Normalizer transformer
     - Added weighted predictions to KNN estimators
+    - Added HTML Stripper transformer
 
 - 0.0.6-alpha
     - Added Gradient Boost regressor